# ext/bcmath: bcpow() performance improvement