@jorgsowa 2024-09-08
Thank you for the review. I changed everything mentioned and tested using `hyperfine`. The difference depending on the number and exponent shape is between 1.05 and 1.08 faster than the previous solution. I attached the results into the description.